+++ .eslintignore
@@ -1,0 +1,1 @@
+cypress/e2e/*
